requirements.txt
grad_cam==1.5.4
kornia==0.7.3
matplotlib==3.10.0
monai==1.3.2
numpy==2.2.3
omegaconf==2.3.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.1.0
pytorch_lightning==2.1.1
PyYAML==6.0.2
rasterio==1.3.10
Requests==2.32.3
rich==13.9.4
segmentation_models_pytorch==0.3.4
skimage==0.0
torch==2.4.0
torchmetrics==1.6.0
torchvision==0.19.0
tqdm==4.66.5
